cars-client/pyproject.toml
[project]
name = "cars-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.57.1",
    "mcp>=1.11.0",
    "python-dotenv>=1.1.1",
]
